refactor/chain: move create_link_descriptor fn
