+++ src/chain/mod.rs
@@ -73,9 +73,9 @@ pub mod node_block;
 mod block_identifier;
 
 pub use chain::block::Block;
-pub use chain::block_identifier::BlockIdentifier;
+pub use chain::block_identifier::{BlockIdentifier, create_link_descriptor};
 pub use chain::data_chain::DataChain;
-pub use chain::node_block::{NodeBlock, Proof, create_link_descriptor};
+pub use chain::node_block::{NodeBlock, Proof};
 use std::fmt::Write;
 
 fn debug_bytes<V: AsRef<[u8]>>(input: V) -> String {
