+++ src/chain/node_block.rs
@@ -15,29 +15,12 @@
 // Please review the Licences for the specific language governing permissions and limitations
 // relating to use of the SAFE Network Software.
 
-use chain::block_identifier::{BlockIdentifier, LinkDescriptor};
+use chain::block_identifier::BlockIdentifier;
 use error::Error;
-use itertools::Itertools;
 use maidsafe_utilities::serialisation;
 use rust_sodium::crypto::sign::{self, PublicKey, SecretKey, Signature};
 use std::fmt::{self, Debug, Formatter};
 use super::debug_bytes;
-use tiny_keccak::Keccak;
-
-/// Returns a link descriptor with the hash of the group members, or `None` if `group` is empty.
-pub fn create_link_descriptor(group: &[PublicKey]) -> Option<LinkDescriptor> {
-    if group.is_empty() {
-        None
-    } else {
-        let mut sha3 = Keccak::new_sha3_256();
-        for key_bytes in group.iter().map(|key| &key.0).sorted() {
-            sha3.update(key_bytes);
-        }
-        let mut res = [0u8; 32];
-        sha3.finalize(&mut res);
-        Some(res)
-    }
-}
 
 /// Proof as provided by a close group member
 #[derive(RustcEncodable, RustcDecodable, PartialOrd, Ord, PartialEq, Eq, Clone)]
